+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Add Windows binary for `groff`. (Since it cannot be installed on Windows natively)
 * Add GUI and logic to get man pages for a command and display the man pages as html.
 * Use a Tab to toggle between comment and man pages.
+* Add html man pages for core utils
+* Support Command Templating and copy to clipboard.
 
 ### CHANGED
 ### DEBUGGED